// client/src/api/model/motivation.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { MotivationType } from './motivationType';

/**
 * A motivation of a specific type
 */
export interface Motivation {
  /** The motivation type */
  type: MotivationType;
  /** Description of the motivation */
  description: string;
}