// client/src/api/model/motivationType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * The type of Motivation
 */
export type MotivationType = typeof MotivationType[keyof typeof MotivationType];


export const MotivationType = {
  Strength: 'Strength',
  Flaw: 'Flaw',
  Desire: 'Desire',
  Fear: 'Fear',
} as const;